修改删除，规范部分代码

+++ src/main/java/com/news/service/impl/CommentServiceImpl.java
@@ -6,6 +6,7 @@ import com.news.service.CommentService;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -46,7 +47,19 @@ public class CommentServiceImpl extends ServiceImpl<CommentMapper, Comment> impl
     }
 
     @Override
-    public boolean batch_delete(List<Integer> idList) {
+    public boolean batch_delete(String deleteIds) {
+        if (deleteIds == null || "".equals(deleteIds)) {
+            return false;
+        }
+        String[] ids = deleteIds.split(",");
+        List<Integer> idList = new ArrayList<>();
+        for (String id : ids) {
+            try {
+                idList.add(Integer.valueOf(id));
+            } catch (NumberFormatException e) {
+                return false;
+            }
+        }
         return commentMapper.deleteBatchIds(idList)>0;
     }
